Move ffprobe binaries
The relative path from the main electron process to the ffprobe-bin folder was different after webpacking which caused the binary to not be found after webpacking. This change ensures that the relative path is always the same by keeping the ffprobe-bin folder at the same level as the electron main process. This is achieved by copying the ffprobe-bin folder into the build directory during the build process